Refactor sysl main package
