+++ sysl2/sysl/exit.go
@@ -1,0 +1,12 @@
+package main
+
+import "fmt"
+
+type exit struct {
+	code    int
+	message string
+}
+
+func exitf(code int, format string, args ...interface{}) exit {
+	return exit{code, fmt.Sprintf(format, args...)}
+}
